backend/requirements.txt
# FastAPI and Web Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Database
sqlmodel>=0.0.14
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.12.0

# AI and MCP
fastmcp>=0.1.0
openai>=1.0.0


# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0

# Testing (dev dependencies)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.27.0,<0.28.0
pytest-cov>=4.1.0
